auth/requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.0
passlib==1.7.4
pip==22.3.1
pluggy==1.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pytest==7.2.1
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.1
starlette==0.22.0
trio==0.22.0
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
wheel==0.38.4
